docker-app/requirements.txt
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==1.10.3
Flask-SocketIO==5.0.1
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-build==3.0.1
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.35.0
gunicorn==20.0.4
honcho==1.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.13.0
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.1
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
wget==3.2
zope.event==4.5.0
zope.interface==5.2.0
